chore: Split HabitsTable in multiple components
