+++ src/app/habits/components/habits-table.tsx
@@ -2,124 +2,98 @@ import { useEffect, useState } from "react";
 import { Habit, Mark, HabitGroup } from "@/app/habits/types";
 import { cn } from "@/lib/utils";
 import React from "react";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { MoreVertical, Pencil, Trash2 } from "lucide-react";
-import { Button } from "@/components/ui/button";
+import { useQuery } from "@tanstack/react-query";
+import { EditableHabitName } from "@/app/habits/components/editable-habit-name";
 
-interface EditableHabitNameProps {
+interface DayCellProps {
+  dayIndex: number | null;
+  isWeekend: boolean;
+}
+
+const DayCell = ({ dayIndex, isWeekend }: DayCellProps) => (
+  <div
+    className={cn(
+      dayIndex !== null ? "bg-white" : "bg-neutral-200",
+      "p-2 text-center text-sm whitespace-nowrap",
+      isWeekend ? "font-bold" : "",
+    )}
+  >
+    {dayIndex !== null && dayIndex + 1}
+  </div>
+);
+
+interface HabitRowProps {
   habit: Habit;
+  days: (number | null)[];
+  onToggleMark: (habit: Habit, day: number) => void;
   onDelete: (habitId: number) => void;
   onUpdate: (habit: Habit) => void;
 }
 
-const EditableHabitName = ({ habit, onDelete, onUpdate }: EditableHabitNameProps) => {
-  const [isEditing, setIsEditing] = useState(false);
-  const [showMenu, setShowMenu] = useState(false);
-  const [editedName, setEditedName] = useState(habit.name);
-
-  const inputRef = React.useRef<HTMLInputElement>(null);
-  const menuRef = React.useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (isEditing) {
-      inputRef?.current?.focus();
-    }
-  }, [isEditing]);
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(event.target as Node)) {
-        setShowMenu(false);
-      }
-    };
-
-    if (showMenu) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [showMenu]);
-
-  const handleEdit = () => {
-    setIsEditing(true);
-    setShowMenu(false);
-  };
-
-  const handleDelete = () => {
-    onDelete(habit.id);
-    setShowMenu(false);
-  };
-
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setEditedName(event.target.value);
-  };
-
-  const handleBlur = () => {
-    if (editedName !== habit.name) {
-      onUpdate({ ...habit, name: editedName });
-    }
-    setIsEditing(false);
-  };
-
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === "Enter") {
-      event.currentTarget.blur();
-    }
-  };
-
-  const toggleMenu = () => setShowMenu(!showMenu);
-
-  if (isEditing) {
-    return (
-      <input
-        className="w-full bg-transparent outline-none"
-        value={editedName}
-        onChange={handleInputChange}
-        onBlur={handleBlur}
-        onKeyDown={handleKeyDown}
-        ref={inputRef}
+const HabitRow = ({ habit, days, onToggleMark, onDelete, onUpdate }: HabitRowProps) => (
+  <React.Fragment>
+    <div className="bg-white p-2 text-sm truncate">
+      <EditableHabitName
+        habit={habit}
+        onDelete={onDelete}
+        onUpdate={onUpdate}
       />
-    );
-  }
-
-  return (
-    <div className="flex items-center justify-between h-full">
-      <span className="flex-1 truncate">{habit.name}</span>
-      <div className="relative flex items-center -mr-2">
-        <Button
-          variant="ghost"
-          size="icon"
-          className="h-6 w-6 p-0"
-          onClick={toggleMenu}
-        >
-          <MoreVertical className="h-4 w-4" />
-        </Button>
-        {showMenu && (
-          <div ref={menuRef} className="fixed transform -translate-y-1 min-w-[120px] rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 z-10">
-            <div className="py-1">
-              <button
-                className="flex items-center w-full px-3 py-1.5 text-sm text-gray-700 hover:bg-gray-100"
-                onClick={handleEdit}
-              >
-                <Pencil className="h-4 w-4 mr-2" />
-                Edit
-              </button>
-              <button
-                className="flex items-center w-full px-3 py-1.5 text-sm text-red-600 hover:bg-gray-100"
-                onClick={handleDelete}
-              >
-                <Trash2 className="h-4 w-4 mr-2" />
-                Delete
-              </button>
-            </div>
-          </div>
-        )}
-      </div>
     </div>
-  );
-};
+    {days.map((dayIndex, i) => (
+      <div
+        key={`${habit.name}-${i}`}
+        className={cn(
+          dayIndex !== null
+            ? "bg-white hover:bg-neutral-50"
+            : "bg-neutral-200",
+          "p-2 text-center text-sm",
+        )}
+        onClick={() =>
+          dayIndex !== null && onToggleMark(habit, dayIndex)
+        }
+        role={dayIndex !== null ? "button" : undefined}
+      >
+        {dayIndex !== null && habit.marks[dayIndex]}
+      </div>
+    ))}
+  </React.Fragment>
+);
+
+interface NewHabitRowProps {
+  newHabit: Habit;
+  days: (number | null)[];
+  onInputChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  onBlur: () => void;
+  inputRef: React.RefObject<HTMLInputElement | null>;
+}
+
+const NewHabitRow = ({ newHabit, days, onInputChange, onBlur, inputRef }: NewHabitRowProps) => (
+  <>
+    <input
+      className="bg-white p-2 text-sm truncate"
+      onChange={onInputChange}
+      onBlur={onBlur}
+      onKeyDown={(event) => {
+        if (event.key === "Enter") {
+          event.currentTarget.blur();
+        }
+      }}
+      ref={inputRef}
+    />
+    {days.map((dayIndex, i) => (
+      <div
+        key={`${newHabit.name}-${i}`}
+        className={cn(
+          dayIndex !== null
+            ? "bg-white hover:bg-neutral-50"
+            : "bg-neutral-200",
+          "p-2 text-center text-sm",
+        )}
+        role={dayIndex !== null ? "button" : undefined}
+      />
+    ))}
+  </>
+);
 
 interface HabitsTableProps {
   currentWeek: number;
@@ -137,8 +111,7 @@ export const HabitsTable = ({
   const [isDesktop, setIsDesktop] = useState<boolean | undefined>(undefined);
   const [newHabit, setNewHabit] = useState<Habit | null>(null);
   const newHabitInputRef = React.useRef<HTMLInputElement>(null);
-  const queryClient = useQueryClient();
-  const { data, isLoading } = useQuery<HabitGroup>({
+  const { data, isLoading, refetch } = useQuery<HabitGroup>({
     queryKey: ["habit-groups", month, year],
     queryFn: async () => {
       const params = new URLSearchParams({
@@ -180,9 +153,7 @@ export const HabitsTable = ({
       },
       body: JSON.stringify({ marks: newMarks }),
     }).then(() =>
-      queryClient.invalidateQueries({
-        queryKey: ["habit-groups", month, year],
-      }),
+      refetch(),
     );
   };
 
@@ -244,10 +215,7 @@ export const HabitsTable = ({
         year: year,
       }),
     }).then(() =>
-      queryClient
-        .invalidateQueries({
-          queryKey: ["habit-groups", month, year],
-        })
+      refetch()
         .then(() => {
           setNewHabit(null);
         }),
@@ -258,9 +226,7 @@ export const HabitsTable = ({
     await fetch(`/api/habits/${habitId}`, {
       method: "DELETE",
     });
-    queryClient.invalidateQueries({
-      queryKey: ["habit-groups", month, year],
-    });
+    refetch();
   };
 
   const handleUpdateHabit = async (updatedHabit: Habit) => {
@@ -271,9 +237,7 @@ export const HabitsTable = ({
       },
       body: JSON.stringify({ name: updatedHabit.name }),
     });
-    queryClient.invalidateQueries({
-      queryKey: ["habit-groups", month, year],
-    });
+    refetch();
   };
 
   useEffect(() => {
@@ -299,71 +263,30 @@ export const HabitsTable = ({
               <button onClick={handleAddHabit}>+</button>
             </div>
             {days.map((dayIndex, i) => (
-              <div
+              <DayCell
                 key={i}
-                className={cn(
-                  dayIndex !== null ? "bg-white" : "bg-neutral-200",
-                  "p-2 text-center text-sm whitespace-nowrap",
-                  isWeekend(dayIndex) ? "font-bold" : "",
-                )}
-              >
-                {dayIndex !== null && dayIndex + 1}
-              </div>
+                dayIndex={dayIndex}
+                isWeekend={isWeekend(dayIndex)}
+              />
             ))}
             {data?.habits?.map((habit) => (
-              <React.Fragment key={habit.id}>
-                <div className="bg-white p-2 text-sm truncate">
-                  <EditableHabitName
-                    habit={habit}
-                    onDelete={handleDeleteHabit}
-                    onUpdate={handleUpdateHabit}
-                  />
-                </div>
-                {days.map((dayIndex, i) => (
-                  <div
-                    key={`${habit.name}-${i}`}
-                    className={cn(
-                      dayIndex !== null
-                        ? "bg-white hover:bg-neutral-50"
-                        : "bg-neutral-200",
-                      "p-2 text-center text-sm",
-                    )}
-                    onClick={() =>
-                      dayIndex !== null && toggleMark(habit, dayIndex)
-                    }
-                    role={dayIndex !== null ? "button" : undefined}
-                  >
-                    {dayIndex !== null && habit.marks[dayIndex]}
-                  </div>
-                ))}
-              </React.Fragment>
+              <HabitRow
+                key={habit.id}
+                habit={habit}
+                days={days}
+                onToggleMark={toggleMark}
+                onDelete={handleDeleteHabit}
+                onUpdate={handleUpdateHabit}
+              />
             ))}
             {newHabit && (
-              <>
-                <input
-                  className="bg-white p-2 text-sm truncate"
-                  onChange={handleInputChange}
-                  onBlur={createNewHabit}
-                  onKeyDown={(event) => {
-                    if (event.key === "Enter") {
-                      event.currentTarget.blur();
-                    }
-                  }}
-                  ref={newHabitInputRef}
-                />
-                {days.map((dayIndex, i) => (
-                  <div
-                    key={`${newHabit.name}-${i}`}
-                    className={cn(
-                      dayIndex !== null
-                        ? "bg-white hover:bg-neutral-50"
-                        : "bg-neutral-200",
-                      "p-2 text-center text-sm",
-                    )}
-                    role={dayIndex !== null ? "button" : undefined}
-                  />
-                ))}
-              </>
+              <NewHabitRow
+                newHabit={newHabit}
+                days={days}
+                onInputChange={handleInputChange}
+                onBlur={createNewHabit}
+                inputRef={newHabitInputRef}
+              />
             )}
           </>
         )}
